Changed absolute refactory period on randon stim input for genesis from 5ms to 0.1 ms Minor updates to gen mappings
git-svn-id: svn://neuroconstruct.org:3999/neuroConstruct/trunk@1077 cb8a3d95-5132-0410-b650-8d485996f008

+++ src/ucl/physiol/neuroconstruct/simulation/SimulationData.java
@@ -634,7 +634,7 @@ public class SimulationData
                     
                     
                     double nextVal = Double.parseDouble(sp[0]);
-                    int cellNum = Integer.parseInt(sp[1]);
+                    int cellNum = (int)Float.parseFloat(sp[1]); // in case it's 0.0, 1.0 etc.
                     //logger.logComment(cellNum+": "+nextVal, true);
                     if (!tempList.containsKey(cellNum))
                         tempList.put(cellNum, new ArrayList<Double>(suggestedInitCapData));
